requirements.txt
streamlit==1.31.1
numpy==1.26.4
Pillow==9.4.0
tensorflow==2.16.1
keras==3.1.1
zipp==3.17.0